Revert moving CLI to lib
